Avoid cyclic dependency by merging callers.py into hooks.py
The two modules have a type-level cyclic dependency:
hooks.py uses _multicall and _legacymulticall from callers.py.
callers.py uses HookImpl from hooks.py.

This suggests that they're better off combined.

The _Result functionality from callers.py is independent and general, so
move it to its own module _result.py.

Backward compatibility: I did not find any external project which
imports directly from pluggy.callers. It exposes two potential import
candidates:
- HookCallError: exposed from the top-level, so no reason to reach
  inside.
- _Result: it's prefixed and hidden, so no guarantees about that.
