+++ docs/api_reference.rst
@@ -9,8 +9,8 @@ Api Reference
     :show-inheritance:
 
 
-.. automethod:: pluggy.callers._Result.get_result
+.. automethod:: pluggy._result._Result.get_result
 
-.. automethod:: pluggy.callers._Result.force_result
+.. automethod:: pluggy._result._Result.force_result
 
 .. automethod:: pluggy.hooks._HookCaller.call_extra
