+++ testing/test_multicall.py
@@ -1,7 +1,6 @@
 import pytest
 from pluggy import HookCallError, HookspecMarker, HookimplMarker
-from pluggy.hooks import HookImpl
-from pluggy.callers import _multicall, _legacymulticall
+from pluggy.hooks import HookImpl, _multicall, _legacymulticall
 
 
 hookspec = HookspecMarker("example")
